Face/requirements.txt
Flask==1.1.1
h5py==2.8.0
Keras==2.3.1
matplotlib==3.2.1
numpy==1.18.1
paho-mqtt==1.5.0
pandas==1.0.3
Pillow==7.0.0
progress==1.5
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorflow-gpu==1.14.0
opencv-python
